movie_backend/requirements.txt
anyio==4.9.0
asgiref==3.9.1
asyncclick==8.1.8.0
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
libtorrent==2.0.11
python-dotenv==1.1.1
requests==2.32.4
sniffio==1.3.1
sqlparse==0.5.3
torrentp==0.2.4
urllib3==2.5.0
